requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==2.6.0
tox==2.3.1
Sphinx==1.6.6
cryptography==1.7
PyYAML==3.12
pytest==3.3.1
pytest-runner==3.0
pytest-cov==2.5.1
pluggy<0.7,>=0.5
codecov==2.0.10
coverage>=3.7.1
mock==2.0.0
six==1.11.0
python-box==3.1.1
